Респавн ч9, устранены недочеты

+++ xrGame/xrServerRespawnManager.h
@@ -7,10 +7,10 @@ public:
 	static void CheckRespawnObjects();
 
 	// В xrServer::Process_event_destroy
-	static int DestroyRespawnID(u16 id);
+	static u16 DestroyRespawnID(u16 id);
 
 	// В xrServer::Process_event_destroy
-	static int GetRespawnObjectID(u16 id);
+	static u16 GetRespawnObjectID(u16 id);
 
 	// добавить объект в респавн xrServer::SLS_Default()
 	static void AddObject(shared_str& pSection, shared_str &pCustomData, u16 pID, int pTimeRespawn, Fvector& XYZ);
